[http] Support booting images larger than 2GB over http on 32-bit builds
The start member of http_request_range has been widened to
64 bit.

The len parameter of the blktrans_xferbuf_realloc has been
widened (which records the length of the image) plus all
intermediate functions that lead to the call.

Tested with an 32GB Windows 7 64-bit image: it boots until
it fails to find the root device (which is expected)

+++ src/include/ipxe/blocktrans.h
@@ -33,6 +33,6 @@ struct block_translator {
 };
 
 extern int block_translate ( struct interface *block,
-			     userptr_t buffer, size_t size );
+			     userptr_t buffer, uint64_t size );
 
 #endif /* _IPXE_BLOCKTRANS_H */
